shuffle

+++ src/main/java/top/ezttf/graduation/controller/spark/Temp.java
@@ -17,5 +17,5 @@ public class Temp implements Serializable {
 
     private long time;
     private long count;
-    private long random;
+    private double random;
 }
